bot-service/requirements.txt
requests
Flask-Pydantic
Flask-SQLAlchemy
clickhouse-connect
python-dotenv
python-dateutil
apscheduler
telebot